# src/swagger_server/models/inference_result.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401
from swagger_server.models.base_model_ import Model
from swagger_server import util

#not visible from init? https://github.com/swagger-api/swagger-codegen/issues/10193
from swagger_server.models.inference_feature import InferenceFeature

class InferenceResult(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, patient_id: int=None, model_id: int=None, features: List[InferenceFeature]=None):  # noqa: E501
        """InferenceResult - a model defined in Swagger

        :param patient_id: The patient_id of this InferenceResult.  # noqa: E501
        :type patient_id: int
        :param model_id: The model_id of this InferenceResult.  # noqa: E501
        :type model_id: int
        :param features: The features of this InferenceResult.  # noqa: E501
        :type features: List[InferenceFeature]
        """
        self.swagger_types = {
            'patient_id': int,
            'model_id': int,
            'features': List[InferenceFeature]
        }

        self.attribute_map = {
            'patient_id': 'patientId',
            'model_id': 'modelId',
            'features': 'features'
        }

        self._patient_id = patient_id
        self._model_id = model_id
        self._features = features

    @classmethod
    def from_dict(cls, dikt) -> 'InferenceResult':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The InferenceResult of this InferenceResult.  # noqa: E501
        :rtype: InferenceResult
        """
        return util.deserialize_model(dikt, cls)

    @property
    def patient_id(self) -> int:
        """Gets the patient_id of this InferenceResult.

        Patients inique identifier is derived from Patient record signature  # noqa: E501

        :return: The patient_id of this InferenceResult.
        :rtype: int
        """
        return self._patient_id

    @patient_id.setter
    def patient_id(self, patient_id: int):
        """Sets the patient_id of this InferenceResult.

        Patients inique identifier is derived from Patient record signature  # noqa: E501

        :param patient_id: The patient_id of this InferenceResult.
        :type patient_id: int
        """

        self._patient_id = patient_id

    @property
    def model_id(self) -> int:
        """Gets the model_id of this InferenceResult.

        Model's inique identifier  # noqa: E501

        :return: The model_id of this InferenceResult.
        :rtype: int
        """
        return self._model_id

    @model_id.setter
    def model_id(self, model_id: int):
        """Sets the model_id of this InferenceResult.

        Model's inique identifier  # noqa: E501

        :param model_id: The model_id of this InferenceResult.
        :type model_id: int
        """

        self._model_id = model_id

    @property
    def features(self) -> List[InferenceFeature]:
        """Gets the features of this InferenceResult.


        :return: The features of this InferenceResult.
        :rtype: List[InferenceFeature]
        """
        return self._features

    @features.setter
    def features(self, features: List[InferenceFeature]):
        """Sets the features of this InferenceResult.


        :param features: The features of this InferenceResult.
        :type features: List[InferenceFeature]
        """

        self._features = features