# src/swagger_server/models/troponin.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Troponin(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, value: int=None, time: int=None):  # noqa: E501
        """Troponin - a model defined in Swagger

        :param value: The value of this Troponin.  # noqa: E501
        :type value: int
        :param time: The time of this Troponin.  # noqa: E501
        :type time: int
        """
        self.swagger_types = {
            'value': int,
            'time': int
        }

        self.attribute_map = {
            'value': 'value',
            'time': 'time'
        }

        self._value = value
        self._time = time

    @classmethod
    def from_dict(cls, dikt) -> 'Troponin':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Troponin of this Troponin.  # noqa: E501
        :rtype: Troponin
        """
        return util.deserialize_model(dikt, cls)

    @property
    def value(self) -> int:
        """Gets the value of this Troponin.


        :return: The value of this Troponin.
        :rtype: int
        """
        return self._value

    @value.setter
    def value(self, value: int):
        """Sets the value of this Troponin.


        :param value: The value of this Troponin.
        :type value: int
        """

        self._value = value

    @property
    def time(self) -> int:
        """Gets the time of this Troponin.

        as day milliseconds  # noqa: E501

        :return: The time of this Troponin.
        :rtype: int
        """
        return self._time

    @time.setter
    def time(self, time: int):
        """Sets the time of this Troponin.

        as day milliseconds  # noqa: E501

        :param time: The time of this Troponin.
        :type time: int
        """

        self._time = time